@@ -1,5 +1,9 @@
 <?php
 
+use App\Http\Controllers\IndexController;
+use App\Http\Controllers\PrintingController;
+use App\Http\Controllers\Auth\LoginController;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -11,23 +15,23 @@
 |
  */
 
-Route::get('/', 'IndexController@index')->name('home');
-Route::get('/printings', 'PrintingController@index');
-Route::get('/printings/admin', 'PrintingController@admin');
-Route::get('/printings/user/{user}', 'PrintingController@user');
-Route::get('/printings/jobid/{jobid}', 'PrintingController@jobid');
-Route::get('/printings/{printer}', 'PrintingController@printer');
+Route::get('/', [IndexController::class, 'index'])->name('home');
+Route::get('/printings', [PrintingController::class, 'index']);
+Route::get('/printings/admin', [PrintingController::class, 'admin']);
+Route::get('/printings/user/{user}', [PrintingController::class, 'user']);
+Route::get('/printings/jobid/{jobid}', [PrintingController::class, 'jobid']);
+Route::get('/printings/{printer}', [PrintingController::class, 'printer']);
 
 # Senha única USP
-Route::get('/senhaunica/login', 'Auth\LoginController@redirectToProvider')->name('login');
-Route::get('/callback', 'Auth\LoginController@handleProviderCallback');
-Route::post('/logout', 'Auth\LoginController@logout');
-Route::get('/logout', 'Auth\LoginController@logout');
+Route::get('/senhaunica/login', [LoginController::class, 'redirectToProvider'])->name('login');
+Route::get('/callback', [LoginController::class, 'handleProviderCallback']);
+Route::post('/logout', [LoginController::class, 'logout']);
+Route::get('/logout', [LoginController::class, 'logout']);
 
-Route::get('/check/{user}/{printer}/{pages}', 'PrintingController@check');
-Route::get('/pages/today/{user}/', 'PrintingController@pagesToday');
+Route::get('/check/{user}/{printer}/{pages}', [PrintingController::class, 'check']);
+Route::get('/pages/today/{user}/', [PrintingController::class, 'pagesToday']);
 
 /* Fila de impressão */
-Route::get('/pendentes', 'PrintingController@pendentes');
-Route::get('/pendentes/{printer}/', 'PrintingController@pendentes');
+Route::get('/pendentes', [PrintingController::class, 'pendentes']);
+Route::get('/pendentes/{printer}/', [PrintingController::class, 'pendentes']);
 
